feat: nginx支持SSL配置；markdown支持语法高亮；优化安全传输策略；解决一为图标补全失效；添加huasen-lib工具库；前台搜索引擎支持可配；支持显示在线人数及强制下线；登录页面支持选择邮箱；支持界面初始化系统管理员；优化界面功能及快捷脚本

+++ huasen-frontend/admin/src/components/content/wrap/sign/WrapSign.vue
@@ -6,27 +6,31 @@
  * @Description: 登录面板
 -->
 <template>
-  <HsDialog title="" :fullscreen="true" :show-close="false" :close-on-press-escape="false" :visible.sync="showWrapSign">
+  <HsDialog v-if="showWrapSign" title="" :fullscreen="true" :show-close="false" :close-on-press-escape="false" :visible.sync="showWrapSign">
     <div class="sign shadow-md rounded-lg">
       <div class="banner">
         <img src="~@/assets/img/sign/default.svg" />
       </div>
       <div class="tab">
         <ul>
-          <li>登录后台管理系统</li>
+          <li>{{ isExistManege ? '登录后台管理系统' : '请添加系统管理员' }}</li>
         </ul>
         <div class="content">
           <span class="panel">
             <section class="login">
-              <el-form :model="submitForm" :rules="rules" ref="loginForm" status-icon>
-                <el-form-item prop="id">
-                  <el-input v-model="submitForm.id" placeholder="请输入邮箱"> </el-input>
+              <el-form :model="submitForm" :rules="rules" ref="loginForm">
+                <el-form-item class="form__id" prop="id">
+                  <el-input v-model="submitForm.mail" @input="setId" placeholder="请输入邮箱">
+                    <el-select v-model="submitForm.mailServer" @change="setId" slot="append">
+                      <el-option v-for="(item, index) in mailOptions" :key="index" :label="item.label" :value="item.value"></el-option>
+                    </el-select>
+                  </el-input>
                 </el-form-item>
-                <el-form-item prop="password">
+                <el-form-item class="form__password" prop="password">
                   <el-input type="password" v-model="submitForm.password" :show-password="true" autocomplete="off" placeholder="请输入密码" @keyup.enter.native="login"></el-input>
                 </el-form-item>
               </el-form>
-              <div class="btn" @click="login">登 录</div>
+              <div class="btn" @click="login">{{ isExistManege ? '登 录' : '添 加' }}</div>
             </section>
           </span>
         </div>
@@ -36,24 +40,62 @@
 </template>
 <script>
 import { mapState, mapMutations } from 'vuex';
-import { getElementFormValidator } from '@/plugin/strategy.js';
+import { Validator } from 'huasen-lib';
+const validator = new Validator();
+const getElementFormValidator = validator.getElementFormValidator.bind(validator);
 import HsDialog from '@/components/common/dialog/Dialog.vue';
 export default {
   name: 'WrapSign',
   components: { HsDialog },
   data() {
     return {
+      // 判断管理员是否存在
+      isExistManege: false,
       // 提交数据
       submitForm: {
         id: '',
         password: '',
+        mail: '',
+        mailServer: '@qq.com',
       },
 
       // 校验的规则
       rules: {
-        id: [{ validator: getElementFormValidator(['isNoEmpty::必填项', 'minLength:5::长度小于5', 'maxLength:20::长度大于20', 'isEmail::请输入邮箱']), trigger: 'blur' }],
-        password: [{ validator: getElementFormValidator(['isNoEmpty::必填项', 'minLength:5::长度小于5', 'maxLength:20::长度大于20']), trigger: 'blur' }],
+        id: [{ validator: getElementFormValidator(['isNonEmpty::必填项', 'minLength:5::长度小于5', 'maxLength:20::长度大于20', 'isEmail::请输入邮箱']) }],
+        password: [{ validator: getElementFormValidator(['isNonEmpty::必填项', 'minLength:5::长度小于5', 'maxLength:20::长度大于20']) }],
       },
+
+      // 邮箱选项
+      mailOptions: [
+        {
+          label: '@qq.com',
+          value: '@qq.com',
+        },
+        {
+          label: '@163.com',
+          value: '@163.com',
+        },
+        {
+          label: '@gmail.com',
+          value: '@gmail.com',
+        },
+        {
+          label: '@139.com',
+          value: '@139.com',
+        },
+        {
+          label: '@126.com',
+          value: '@126.com',
+        },
+        {
+          label: '@outlook.com',
+          value: '@outlook.com',
+        },
+        {
+          label: '自定义邮箱',
+          value: '',
+        },
+      ],
     };
   },
   computed: {
@@ -68,22 +110,65 @@ export default {
       },
     },
   },
+  watch: {
+    showWrapSign(val) {
+      if (val) {
+        this.existMange();
+      }
+    },
+  },
+  async created() {
+    await this.existMange();
+  },
   methods: {
     ...mapMutations(['commitAll']),
+
+    /**
+     * 判断管理员是否存在
+     */
+    async existMange() {
+      const res = await this.API.manage.exist({}, { notify: false });
+      this.isExistManege = !!res.data;
+    },
+
+    /**
+     * 设置邮箱地址
+     */
+    setId() {
+      this.submitForm.id = this.submitForm.mail + this.submitForm.mailServer;
+    },
+
     // 登录处理
     login() {
       this.$refs.loginForm.validate(valid => {
         if (valid) {
-          this.API.login(this.submitForm, {
-            secret: true,
-          }).then(res => {
-            this.STORAGE.setItem(this.CONSTANT.localManage, res.data);
-            // 关闭登录面板
-            this.commitAll({
-              showWrapSign: false,
-            });
-            location.reload();
-          });
+          if (this.isExistManege) {
+            this.API.manage
+              .login(this.submitForm, {
+                secret: 'rsa',
+              })
+              .then(res => {
+                this.STORAGE.setItem(this.CONSTANT.localManage, res.data);
+                // 关闭登录面板
+                this.commitAll({
+                  showWrapSign: false,
+                });
+                location.reload();
+              });
+          } else {
+            this.$confirm('该操作将为系统初始化管理员，请务必确保输入信息的正确性，一旦初始化后将无法再次更改！', '添加系统管理员', {
+              confirmButtonText: '确定',
+              cancelButtonText: '取消',
+              type: 'warning',
+            })
+              .then(async () => {
+                await this.API.manage.init(this.submitForm, {
+                  secret: 'rsa',
+                });
+                this.isExistManege = true;
+              })
+              .catch(error => {});
+          }
         }
       });
     },
@@ -130,15 +215,25 @@ export default {
         box-sizing: border-box;
         cursor: pointer;
       }
-      ::v-deep .el-form {
-        .el-form-item__content {
-          input {
-            border-radius: 0;
-            border: none;
-            padding: 0;
-            border-bottom: 1px solid var(--gray-400);
+    }
+  }
+}
+</style>
+<style lang="scss">
+.sign {
+  .el-form {
+    .el-form-item {
+      &.form__id {
+        .el-input {
+          width: 100%;
+          .el-input-group__append {
+            .el-select {
+              width: 148px;
+            }
           }
         }
+      }
+      &.form__password {
       }
     }
   }
